Get matrix with ratios/educts/products for ropa
New modules ropa and prepare_ropa to for chemical sink and source
analysis. Current routines in prepare_ropa get list of species names and
reactions and a matrix with branching ratios and species names for
educts and products, respectively. The general workflow is designed in
the main script 'ropa.jl'.
Performance updates in DSMACCplot.jl allowing to push the directories of
self-made modules to LOAD_PATH independently of the current directory
position.
Further functions in ropa.jl to prepare the flux analysis.