Revised folder structure, documentation, fixes
Renaming folder IO to DATA, now only holding the database file with the
species names/formulas in different languages (MCM, SMILES, InChI,
GECKO-A). Moving the input file with the plot specifications for
DSMACCplot.jl to the main folder.
In DSMACCplot.jl, specifying the database file as
normpath(joinpath(Base.source_dir(),"DATA/MCMv331species.db"))
to be able to call the script either from the DSMACC main folder or the
DSMACCanalysis folder.
In module NCload, specifying the folder path holding the python library
to read netCDF files with joinpath(Base.source_dir(),"py.lib") to be
able to call the script either from the DSMACC main folder or the
DSMACCanalysis folder.
In module jlplot, changing the default folder of the input file with the
plot specifications from AnalysisTools to AnalysisTools/DSMACCanalysis.
Defining default path of output pdf with plots as ./save/results.
Allowing universally whitespaces, commas, and semicolons as separators
in horizontal lists in the plot definition file. This includes changes
in jlmod in function get_scenario to allow whitespace and semicolons in
the input nc file list, get_settings to allow whitespaces and semicolons
in the definition of the cut-off, and function prepare_plots to allow
whitespaces and semicolons in the definitions of scenarios to be
plotted.
New README including a version history.
FIX: In module jlplot in function get_scenario, correct the retrieval of
     scenario labels. Previously all labels would have been overwritten
     with "v3" and "v4".