I/O file handling, fixes
At the end of the preamble in the main script DSMACCplot.jl,  setting a
default directory for the input file in the parent folder above the
repository to allow a clean repository after changes in the input file.
At the beginning of the plotting section, adjusting the file naming of
the output file. The ending ".pdf" is now optional in the second script
argument. A default directory is set to "./save/results" within the
DSMACC repository (git@github.com:pb866/DSMACC-testing.git).
Updated input files and README.
Further changes in module make_plots in repo auxdata
(https://github.com/pb866/auxdata.git).
FIX: 'night' is now set to [-1] in function def_night in jlplot to avoid
     indexing errors in the plotting loop. Further adjustments of night-
     time shading parameters in the plotting loop for line plots.