Redesigning source code structure
Splitting the DSMACCplot script into three parts moving to different
files in a new SRC folder. The preamble is moved to preamble.jl,
functions to load and analyse data are in load.jl and the plotting loop
is in plot.jl.
Moving the jl.mod and py.lib folder to SRC. Adjusting julia modules and
python libraries paths to new folder structure.